Core: Refactor and remove useless files
