+++ src/main/scala/Top/SimTopIO.scala
@@ -1,0 +1,11 @@
+package Top
+
+import chisel3._
+import difftest._
+
+class SimTopIO extends Bundle {
+  val logCtrl = new LogCtrlIO
+  val perfInfo = new PerfInfoIO
+  val uart = new UARTIO
+}
+
